finish implementation of truetype cmap replacer #98
this overwrites the cmap table which is moved to the end of the truetype font file. the new table contains a single windows symbol subtable (3,0) of the format 6 type which maps the character codes in the single range 0 -> glyphcount to the corresponding glyph indices in the font. the new cmap table is then written to the new font file and the header value for length is updated.

this also changes many truetype classes to use the corresponding ushort datatype rather than ints, to save space.

the generated file is displaying correctly in most pdf viewers and passes all tests but in adobe acrobat reader the text is present but invisible. this was not a problem with the previous approach to file generation. there is no log information as to why this might be the case but it seems like the answer must be related to the validity of the overwritten truetype file. we might need to provide an additional macroman cmap subtable in case this is required by acrobat reader.

running the produced font for andada-regular through fontvalidator https://github.com/HinTak/Font-Validator/releases indicates a number of issues that may cause the file to be an invalid font (cannot open with microsoft font viewer). the next step is to compare the errors present in the unmodified andada regular file with the errors in our version. the most likely candidates seem to be:
* os/2: font is a symbol font but panose byte 1, familytype, is not set to latin symbol.
* os/2: a unicode range was indicated in ulunicoderange but the font has no characters in that range.
* os/2: the usfirstcharindex/uslastcharindex is not valid.
* os/2: the font contains a 3,0 cmap but the codepagerange bit 31 is clear.
* os/2: the usbreakchar is not mapped to a glyf.
* head: font checksum is incorrect (this can also be the case for working fonts so seems unlikely to be the cause).

+++ src/UglyToad.PdfPig/Fonts/TrueType/Glyphs/Glyph.cs
@@ -18,7 +18,7 @@
         /// <summary>
         /// An array of the last points of each contour.
         /// </summary>
-        public int[] EndPointsOfContours { get; }
+        public ushort[] EndPointsOfContours { get; }
 
         public GlyphPoint[] Points { get; }
 
@@ -26,7 +26,7 @@
 
         public bool IsEmpty => Points.Length == 0;
 
-        public Glyph(bool isSimple, byte[] instructions, int[] endPointsOfContours, GlyphPoint[] points,
+        public Glyph(bool isSimple, byte[] instructions, ushort[] endPointsOfContours, GlyphPoint[] points,
             PdfRectangle bounds)
         {
             IsSimple = isSimple;
@@ -38,7 +38,7 @@
 
         public static IGlyphDescription Empty(PdfRectangle bounds)
         {
-            return new Glyph(true, new byte[0], new int[0], new GlyphPoint[0], bounds);
+            return new Glyph(true, new byte[0], new ushort[0], new GlyphPoint[0], bounds);
         }
 
         public IGlyphDescription DeepClone()
@@ -46,7 +46,7 @@
             var clonedInstructions = new byte[Instructions.Length];
             Array.Copy(Instructions, clonedInstructions, Instructions.Length);
 
-            var clonedEndPoints = new int[EndPointsOfContours.Length];
+            var clonedEndPoints = new ushort[EndPointsOfContours.Length];
             Array.Copy(EndPointsOfContours, clonedEndPoints, EndPointsOfContours.Length);
 
             var clonedPoints = new GlyphPoint[Points.Length];
@@ -59,7 +59,7 @@
         {
             var newPoints = MergePoints(glyph);
             var newEndpoints = MergeContourEndPoints(glyph);
-            
+
             return new Glyph(false, Instructions, newEndpoints, newPoints, Bounds);
         }
 
@@ -80,11 +80,11 @@
             return newPoints;
         }
 
-        private int[] MergeContourEndPoints(IGlyphDescription glyph)
+        private ushort[] MergeContourEndPoints(IGlyphDescription glyph)
         {
             var destinationLastEndPoint = EndPointsOfContours[EndPointsOfContours.Length - 1] + 1;
 
-            var endPoints = new int[EndPointsOfContours.Length + glyph.EndPointsOfContours.Length];
+            var endPoints = new ushort[EndPointsOfContours.Length + glyph.EndPointsOfContours.Length];
 
             for (var i = 0; i < EndPointsOfContours.Length; i++)
             {
@@ -93,7 +93,7 @@
 
             for (var i = 0; i < glyph.EndPointsOfContours.Length; i++)
             {
-                endPoints[i + EndPointsOfContours.Length] = glyph.EndPointsOfContours[i] + destinationLastEndPoint;
+                endPoints[i + EndPointsOfContours.Length] = (ushort)(glyph.EndPointsOfContours[i] + destinationLastEndPoint);
             }
 
             return endPoints;
